# swagger_api_documentation.yaml
openapi: 3.1.0
info:
  title: NirogPath API
  version: 1.0.0
  description: |
    REST API for the **NirogPath Healthcare & OPD Prescription Hub Platform**.
    A comprehensive healthcare management system supporting Patients, Doctors, Clinic Receptionists, and Superuser Admins.
    
    ## Authentication & Security
    Protected endpoints require a HTTP Bearer JWT token passed in the `Authorization` header:
    `Authorization: Bearer <jwt_token>`
    
    Tokens are obtained by authenticating via `POST /api/auth/login` or `POST /api/auth/register`.
    
    ## System Roles & Access Control Matrix
    - **patient**: Browse doctors/hospitals, book consultation slots, pay token deposit via wallet, track live OPD queue position, view e-prescriptions, log daily medication doses, manage family profiles, and ask prescription follow-up questions.
    - **doctor**: Access doctor queue, advance consultation state (`call-next`, `complete`), broadcast running late warnings, set location & availability schedules, write e-prescriptions, reply to patient prescription queries.
    - **receptionist**: Clinic-wide aggregated metrics across doctors, hospital map uploads, walk-in token creation, and patient transfer management.
    - **admin**: Superuser operations — register doctors, create partner hospitals, manage users, and view hospital analytics.
    
    ## User Story Mapping
    | User Story ID | Role | Description | API Endpoints |
    |---|---|---|---|
...
    
...
paths:
...
  /api/auth/me:
    get:
      summary: Me
      operationId: me_api_auth_me_get
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

    patch:
      summary: Update Profile
      description: Update profile fields for any authenticated user.
      operationId: update_profile_api_auth_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdateIn"
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /api/doctor/bookings/{booking_id}/prescription:
    post:
      summary: Write Prescription
      operationId: write_prescription_api_doctor_bookings__booking_id__prescription_post
      parameters:
        - name: booking_id
          in: path
          required: true
          schema:
            type: string
            title: Booking Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionIn"
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /api/prescriptions/{prescription_id}:
    get:
      summary: Get Prescription
      operationId: get_prescription_api_prescriptions__prescription_id__get
      parameters:
        - name: prescription_id
          in: path
          required: true
          schema:
            type: string
            title: Prescription Id
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /api/uploads/avatars/{filename}:
    get:
      summary: Get Avatar
      description: Serve uploaded avatar files.
      operationId: get_avatar_api_uploads_avatars__filename__get
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
            title: Filename
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /api/auth/change-password:
    post:
      summary: Change Password
      description: Change password for patient. Verifies current password before updating.
      operationId: change_password_api_auth_change_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordIn"
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /api/uploads/avatar:
    post:
      summary: Upload Avatar
      description: Upload a patient avatar image. Max 1 MB, jpg/png/webp only.
      operationId: upload_avatar_api_uploads_avatar_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/Body_upload_avatar_api_uploads_avatar_post"
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/wallet/topup:
    post:
      summary: Wallet Topup
      description: Add money to the patient's wallet. Amount must be one of the allowed values.
      operationId: wallet_topup_api_wallet_topup_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopupIn"
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/wallet/redeem:
    post:
      summary: Wallet Redeem
      description: Redeem a promo code to credit the patient's wallet.
      operationId: wallet_redeem_api_wallet_redeem_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedeemIn"
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /api/prescriptions/{id}/verify:
    get:
      summary: Verify Prescription
      description: PUBLIC — no auth required. Returns minimal prescription validity info.
      operationId: verify_prescription_api_prescriptions__id__verify_get
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            title: Id
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/prescriptions/{id}/print-data:
    get:
      summary: Prescription Print Data
      description: |
        Auth required — patient owner OR writing doctor.
        Returns full prescription with doctor signature, license, and credentials.
      operationId: prescription_print_data_api_prescriptions__id__print_data_get
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            title: Id
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    Body_upload_avatar_api_uploads_avatar_post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
        - file
      title: Body_upload_avatar_api_uploads_avatar_post
...
    ChangePasswordIn:
      properties:
        current_password:
          type: string
          title: Current Password
        new_password:
          type: string
          title: New Password
      type: object
      required:
        - current_password
        - new_password
      title: ChangePasswordIn
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MedicationIn:
      properties:
        name:
          type: string
          minLength: 1
          title: Name
        dose:
          type: string
          minLength: 1
          title: Dose
        frequency:
          type: string
          title: Frequency
        food_instructions:
          type: string
          title: Food Instructions
          default: any
        duration_days:
          type: integer
          maximum: 365.0
          minimum: 1.0
          title: Duration Days
          default: 5
      type: object
      required:
        - name
        - dose
        - frequency
      title: MedicationIn
...
    PrescriptionIn:
      properties:
        medications:
          items:
            $ref: "#/components/schemas/MedicationIn"
          type: array
          title: Medications
        notes:
          anyOf:
            - type: string
            - type: null
          title: Notes
          default: ""
      type: object
      required:
        - medications
      title: PrescriptionIn
...
    ProfileUpdateIn:
      properties:
        name:
          anyOf:
            - type: string
            - type: null
          title: Name
        phone:
          anyOf:
            - type: string
            - type: null
          title: Phone
        dob:
          anyOf:
            - type: string
            - type: null
          title: Dob
        gender:
          anyOf:
            - type: string
            - type: null
          title: Gender
        avatar_url:
          anyOf:
            - type: string
            - type: null
          title: Avatar Url
        preferred_language:
          anyOf:
            - type: string
            - type: null
          title: Preferred Language
      type: object
      title: ProfileUpdateIn
    RedeemIn:
      properties:
        code:
          type: string
          title: Code
      type: object
      required:
        - code
      title: RedeemIn
...
    TopupIn:
      properties:
        amount:
          type: integer
          title: Amount
      type: object
      required:
        - amount
      title: TopupIn
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...